Add two placeholder components for analytics work
